Add files via upload
I share with you a database of electromyographic signals obtained from a group of people (women), in a range of ages between 22 and 28 years with certain characteristics in common, who performed a walking protocol. This database is used in the R code to extract muscle patterns that allow me to identify types of gait limitation, considering only the knee joint.