Update README.md
If you wish to reproduce the results obtained for constraint identification, as seen in the article entitled ‘Mobility Deficit Identification and Compensation through an Artificial Neural Network and Adaptive Controller Design during Gait’, you should consider a series of steps as described below. 

The EMGDATA2.csv database above refers to a walking protocol obtained from a group of women in a specific age range between 22 and 28 years with specific characteristics.